Moved ApiException into the Portable project
This is super lazy, but putting ApiException in RestService.cs means I
don't have to change 230958 project files.

Fixes #126.
